requirements.txt
dropbox==2.0.0
python-gnupg==0.3.6
urllib3==1.7.1
wsgiref==0.1.2
